JQ filtering in client (#20)
Add ability to specify a jq filter that is applied to the response body. Allows extracting only parts of the response, which is useful when the response is verbose and the server is only interested in specific part of the response. An example use case is listing Kubernetes resources where the metadata of each item in the response is very verbose. (See test case for a concrete example.) Uses the node-jq library.