Braekhus client: terminate websocket on ping timeout (#52)
Follows the pattern provided in the ws library documentation
(https://github.com/websockets/ws#how-to-detect-and-close-broken-connections)
to implement a ping timeout that terminates the client websocket if it
has not received a ping from the server within the timeout time.

The server pings every 5 seconds, so the timeout is set to 30 seconds to
start (6 failed pings). When the connection is terminated, the 'close'
handler will be called, which already includes an attempt to retry the
connection with a backoff. Therefore, this change means that if the
client has not received pings from the server in 30s, it will retry the
connection.

The motivation for this change is the issues that a customer is
encountering. The fix has been to restart the container, but maybe
retrying the websocket connection is sufficient. Regardless of whether
this is the underlying issue, this is recommended by the library for
situations in which the client is unaware that the connection is lost,
so we should implement it anyway.

Tested on dev by increasing the ping interval from the server to make
sure the timeout is triggered and the connection reopens.